% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollen.R
\docType{data}
\name{pollen}
\alias{pollen}
\title{Measurements of the various pollen monitors}
\format{
A list of three data frames with 176 (daily), 704 (sixhour)
and 3168 (hourly) rows and 5 variables:
\describe{
  \item{datetime}{the date and time, in UTC}
  \item{date}{the date, in UTC}
  \item{hour}{the hour, in UTC}
  \item{trap}{the pollen monitor, name of the trap}
  \item{conc}{the measured pollen concentration, in [m^(-3)]}
}
}
\usage{
pollen
}
\description{
A list of three data sets containing pollen measurements.
The list elements represent one dataframe each, containing
daily, sixhour and hourly pollen concentrations.
This data set does NOT contain missing values
}
\keyword{datasets}
