% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pollen_full.R
\docType{data}
\name{pollen_full}
\alias{pollen_full}
\title{Measurements of the various pollen monitors}
\format{
A list of three data frames with 344 (daily), 1376 (sixhour)
and 6192 (hourly) rows and 5 variables:
\describe{
  \item{datetime}{the date and time, in UTC}
  \item{date}{the date, in UTC}
  \item{hour}{the hour, in UTC}
  \item{trap}{the pollen monitor, name of the trap}
  \item{conc}{the measured pollen concentration, in [m^(-3)]}
}
}
\usage{
pollen_full
}
\description{
A list of three data sets containing pollen measurements.
The list elements represent one dataframe each, containing
daily, sixhour and hourly pollen concentrations.
This data set contains missing values as NA
}
\keyword{datasets}
